application/requirements-test.txt
pytest
sanic-testing
